hdis_appointment_management/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
mysqlclient
django-mysql==4.4.0
django-cors-headers==3.11.0
pika==1.2.0
requests
python-dateutil